Added thread to automatically force robots and clusters IR update. Rounded values of IR into the database.

+++ database.sql
@@ -9,7 +9,7 @@ USE `ir_viewer`;
 CREATE TABLE `cluster`(
   `cluster_id` INT PRIMARY KEY,
   `area_id` INT NOT NULL,
-  `cluster_ir` DOUBLE NOT NULL DEFAULT '0'
+  `cluster_ir` DOUBLE(5,2) NOT NULL DEFAULT '0'
 );
 
 /* ROBOT */
@@ -17,6 +17,6 @@ CREATE TABLE `cluster`(
 CREATE TABLE `robot`(
   `robot_id` INT PRIMARY KEY,
   `cluster_id` INT NOT NULL,
-  `robot_ir` DOUBLE NOT NULL DEFAULT '0',
+  `robot_ir` DOUBLE(5,2) NOT NULL DEFAULT '0',
   FOREIGN KEY (cluster_id) REFERENCES cluster(`cluster_id`) ON DELETE CASCADE
 );
